+++ R/add_posterior_forecasts.modellist.R
@@ -45,7 +45,7 @@
 add_posterior_forecasts.modellist <- function(object, ..., cores = 1){
 
   if (.use_cluster(object, cores)) {
-    return(.simulate_models_in_parallel(object, add_posterior_forecasts, cores, ...))
+    return(.simulate_models_in_parallel(object, "add_posterior_forecasts", cores, ...))
   }
   
   for (i in 1:length(object)) {